+++ src/mockActivities.js
@@ -1,0 +1,34 @@
+export default [
+  {
+    id: 'bc2266b2-d6c3-11eb-aef5-6d77908bfcbb',
+    published: '2021-06-25T21:16:12.827Z',
+  },
+  {
+    id: 'c6d88ee0-d6c3-11eb-8b76-fb213805e947',
+    published: '2021-06-23T21:16:30.798Z',
+  },
+  {
+    id: '5f5ad4a0-d7b6-11eb-a8bc-1badfdac8742',
+    published: '2021-06-24T02:13:04.874Z',
+  },
+  {
+    id: 'd33057d0-d831-11eb-a954-81c4a13d6973',
+    published: '2021-06-25T16:56:47.309Z',
+  },
+  {
+    id: '69128f70-d841-11eb-9511-0d3f153f12e6',
+    published: '2021-06-26T18:48:21.223Z',
+  },
+  {
+    id: '81a8f6a0-d841-11eb-a4e5-af1950fe4a84',
+    published: '2021-06-27T18:49:02.474Z',
+  },
+  {
+    id: 'b94c20a0-d841-11eb-9efc-3f4590d442ae',
+    published: '2021-06-28T18:50:35.818Z',
+  },
+  {
+    id: 'c0270a70-d841-11eb-9cb2-cfdb30c70966',
+    published: '2021-06-29T18:50:47.319Z',
+  },
+];